feat: cuando un paquete de monitoreo de delay no llegue, se considera que el enlace esta colapsado por lo que asigana un delay de 9999 segundos

+++ DRSIR-DRL-routing-approach-for-SDN/SDNapps_proac/simple_delay.py
@@ -129,7 +129,7 @@ class simple_Delay(app_manager.RyuApp):
             delay = (fwd_delay + re_delay - src_latency - dst_latency)/2
             return max(delay, 0)
         except:
-            return float(0)
+            return float(9999)
 
     def _save_lldp_delay(self, src=0, dst=0, lldpdelay=0):
         try:
